+++ MetaWear/Core/CBUUID.swift
@@ -35,7 +35,7 @@
 
 import CoreBluetooth
 
-
+/// Bluetooth ID's used by MetaWear
 extension CBUUID {
     public static let metaWearService = CBUUID(string: "326A9000-85CB-9195-D9DD-464CFBBAE75A")
     public static let metaWearCommand = CBUUID(string: "326A9001-85CB-9195-D9DD-464CFBBAE75A")