+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Bluetooth State Restoration on MetaWearScanner.shared
+
+### Changed
+- CPP submodule accessed with https instead of git protocol
 
 ## [3.1.4] - 2018-07-23